Documentation Update - README improvements - Docstring fixes - Adding util objects to __init__.py - Collapsing util.py back into client.py
