+++ paramiko_jump/__init__.py
@@ -2,14 +2,22 @@
 
 paramiko_jump: Paramiko + Jump Host + Multi-Factor Authentication
 
+A simple recipe for an easy approach to using paramiko to SSH
+through a "Jump Host", with support for keyboard-interactive
+multi-factor-authentication.
+
 """
 
 from paramiko_jump.client import (
+    DummyAuthHandler,
     SSHJumpClient,
+    jump_host,
     simple_auth_handler,
 )
 
 __all__ = (
+    'DummyAuthHandler',
     'SSHJumpClient',
+    'jump_host',
     'simple_auth_handler',
 )
